Added unit conversion for Visibility

+++ domain/src/main/java/one/mann/domain/model/CurrentWeather.kt
@@ -14,5 +14,5 @@ data class CurrentWeather(
         val windSpeed: Float,
         val windDirection: Int,
         var lastUpdated: Long,
-        val visibility: Int
+        val visibility: Float
 )
